Expose `RELEASE` build env var in the stage scope
The instruction `ARG RELEASE` was only present before the `FROM` line
making it out of scope. This change exposes it to the main image
build stage.

Note that this change makes the `${RELEASE}` env var available in the
runtime of each `RUN` instruction which may cause unintended
side-effects whenever any underlying tooling depends on the value in
this variable.

Ref: https://github.com/pyca/infra/issues/338