Make use of shebang in the build scripts
This change switches over to calling the executable scripts directly
letting the runtime to consult with the shebang and use the
appropriate interpreter.

Old CentOS-based images have `sh` pointing to `bash` while in Debian
it actually points to `sh`. This is a side-effect that the old image
used to rely on and it does not work anymore for manylinux_2_24.

One way would be to change `s/sh/bash` in the dockerfile but I chose
to rely on the existing shebang comments that already point to it but
are ignored when the interpreter is stated explicitly.