Add `manylinux_2_24` images to the build matrix (#345)
* Install `prelink` w/ OS-specific package manager

manylinux1/2010/2014 are CentOS-based but 2_24 runs Debian 9. This
is why this patch switches over to using APT conditionally.

* Make use of shebang in the build scripts

This change switches over to calling the executable scripts directly
letting the runtime to consult with the shebang and use the
appropriate interpreter.

Old CentOS-based images have `sh` pointing to `bash` while in Debian
it actually points to `sh`. This is a side-effect that the old image
used to rely on and it does not work anymore for manylinux_2_24.

One way would be to change `s/sh/bash` in the dockerfile but I chose
to rely on the existing shebang comments that already point to it but
are ignored when the interpreter is stated explicitly.

* Add `manylinux_2_24` images to GHA build matrix

Fixes #339

* Fix bash argument expansion in install_libffi.sh

* Update .github/workflows/build-docker-images.yml

Co-authored-by: Alex Gaynor <alex.gaynor@gmail.com>

Co-authored-by: Paul Kehrer <paul.l.kehrer@gmail.com>
Co-authored-by: Alex Gaynor <alex.gaynor@gmail.com>