don't build libffi, install libffi-dev on aarch64 (#347)
We originally built our own libffi because we built and distributed cffi
wheels on behalf of that project. Since we no longer do this building
the latest libffi is not a requirement and we can use the wheel version
where available (x86_64) or system libffi where wheels are not
currently build (aarch64). To make this work we need to install
libffi-dev on manylinux_2_24. libffi-devel is already installed on
manylinux2014.